+++ geniustopics/preprocessing.py
@@ -1,4 +1,10 @@
 import regex
+from nltk.corpus import stopwords
+from nltk.stem import WordNetLemmatizer
+from nltk.tokenize import word_tokenize, sent_tokenize
+from .data.vocab import vocab_stop_words, vocab_replacements
+
+lemmatizer = WordNetLemmatizer()
 
 
 def remove_new_line(input_string: str):
@@ -38,7 +44,7 @@ def strip_text_w_sq_brackets(input_string: str):
     return regex.sub(pattern, '', input_string)
 
 
-def strip_punctuation(input_string):
+def strip_punctuation(input_string: str):
     """
     Strip punctuation from the input string and return the modified string.
 
@@ -55,8 +61,32 @@ def strip_punctuation(input_string):
     return regex.sub(pattern, '', input_string)
 
 
-def strip_white_space(input_string):
+def strip_white_space(input_string: str):
     pattern = regex.compile(r"\s{2,}")
     if not isinstance(input_string, str):
         return ""
     return regex.sub(pattern, ' ', input_string).strip()
+
+
+def normalize_case(input_string: str):
+    if not isinstance(input_string, str):
+        return ""
+    return input_string.casefold()
+
+
+def strip_stop_words(input_string: str):
+    if not isinstance(input_string, str):
+        return ""
+    input_list = input_string.split(' ')
+    stop_words = stopwords.words('english')
+    stop_word_removed_list = [
+        word for word in input_list if word not in stop_words]
+    stop_word_removed_list = [
+        word for word in stop_word_removed_list if word not in vocab_stop_words]
+    return " ".join(stop_word_removed_list)
+
+
+def lemmatize(input_string: str):
+    if not isinstance(input_string, str):
+        return ""
+    return lemmatizer.lemmatize(input_string)